% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpicker.R
\name{colorpickerInput}
\alias{colorpickerInput}
\title{<Add Title>}
\usage{
colorpickerInput(inputId, defaultColor = "#fff", type = c("sketch",
  "alpha", "block", "chrome", "circle", "compact", "github", "hue",
  "material", "photoshop", "slider", "swatches", "twitter"))
}
\description{
<Add Description>
}
