% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpicker.R
\name{colorpickerInput}
\alias{colorpickerInput}
\title{Create a color picker input}
\usage{
colorpickerInput(inputId, defaultColor = "#ffffff", type = c("sketch",
  "alpha", "block", "chrome", "circle", "compact", "github", "hue",
  "photoshop", "slider", "swatches", "twitter"))
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the chosen
color value. The color value is a hexadecimal string of the form
\code{"#rrggbb"} or \code{"#rrggbbaa"}.}

\item{defaultColor}{An R color as accepted by
\code{\link[grDevices]{col2rgb}} to serve as the initial value of the
input. Defaults to \code{"#ffffff"} (white).}

\item{type}{The lower-case name of the react-color picker type; defaults to
\code{"sketch"}. The following types are available:
\itemize{
  \item \code{"sketch"}
  \item \code{"alpha"}
  \item \code{"block"}
  \item \code{"chrome"}
  \item \code{"circle"}
  \item \code{"compact"}
  \item \code{"github"}
  \item \code{"hue"}
  \item \code{"photoshop"}
  \item \code{"slider"}
  \item \code{"swatches"}
  \item \code{"twitter"}
}}
}
\value{
A color picker widget suitable for inclusion in a Shiny UI.
}
\description{
Creates a color picker input powered by the
\href{https://github.com/casesandberg/react-color}{react-color} JavaScript
library. There are 13 available types of picker, each with a different
appearance and behavior. The picker type is determined by the value of the
\code{type} argument.
}
